% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_newsmap.R
\name{accuracy}
\alias{accuracy}
\title{Evaluate classification accuracy in precision and recall}
\usage{
accuracy(x, y)
}
\arguments{
\item{x}{vercor of predicted classes}

\item{y}{vector of true classes}
}
\description{
Evaluate classification accuracy in precision and recall
}
