% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_newsmap.R
\name{afe}
\alias{afe}
\title{Compute average feature entroy}
\usage{
afe(x, y, smooth = 1)
}
\arguments{
\item{x}{a dfm for features}

\item{y}{a dfm for labels}

\item{smooth}{a numeric value for smoothing to include all the features}
}
\description{
Compute average feature entroy
}
