% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_newsmap_ar}
\alias{data_dictionary_newsmap_ar}
\title{Seed geographical dictionary in Arabic}
\description{
Seed geographical dictionary in Arabic
}
\author{
Dai Yamao \email{daiyamao@scs.kyushu-u.ac.jp}
}
\keyword{data}
