% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_newsmap_de}
\alias{data_dictionary_newsmap_de}
\title{Seed geographical dictionary in German}
\description{
Seed geographical dictionary in German
}
\author{
Stefan Müller \email{mullers@tcd.ie}
}
\keyword{data}
