% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_newsmap_en}
\alias{data_dictionary_newsmap_en}
\title{Seed geographical dictionary in English}
\description{
Seed geographical dictionary in English
}
\author{
Kohei Watanabe \email{watana.kohei@gmail.com}
}
\keyword{data}
