% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_newsmap_es}
\alias{data_dictionary_newsmap_es}
\title{Seed geographical dictionary in Spanish}
\description{
Seed geographical dictionary in Spanish
}
\author{
Dani Madrid-Morales \email{dani.madrid@my.cityu.edu.hk}
}
\keyword{data}
