% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_newsmap_fr}
\alias{data_dictionary_newsmap_fr}
\title{Seed geographical dictionary in French}
\description{
Seed geographical dictionary in French
}
\author{
Claude Grasland \email{claude.grasland@parisgeo.cnrs.fr}
}
\keyword{data}
