% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_newsmap_he}
\alias{data_dictionary_newsmap_he}
\title{Seed geographical dictionary in Hebrew}
\description{
Seed geographical dictionary in Hebrew
}
\author{
Elad Segev \email{eladseg@gmail.com}
}
\keyword{data}
