% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_newsmap_it}
\alias{data_dictionary_newsmap_it}
\title{Seed geographical dictionary in Italian}
\description{
Seed geographical dictionary in Italian
}
\author{
Giuseppe Carteny \email{giuseppe.carteny@unimi.it}
}
\keyword{data}
