% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_newsmap_ja}
\alias{data_dictionary_newsmap_ja}
\title{Seed geographical dictionary in Japanese}
\description{
Seed geographical dictionary in Japanese
}
\author{
Kohei Watanabe \email{watanabe.kohei@gmail.com}
}
\keyword{data}
