% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_newsmap_pt}
\alias{data_dictionary_newsmap_pt}
\title{Seed geographical dictionary in Portuguese}
\description{
Seed geographical dictionary in Portuguese
}
\author{
Barbara Ellynes Zucchi Nobre Silva \email{xx@yyy}
}
\keyword{data}
