% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_newsmap_ru}
\alias{data_dictionary_newsmap_ru}
\title{Seed geographical dictionary in Russian}
\description{
Seed geographical dictionary in Russian
}
\author{
Katerina Tertytchnaya \email{katerina.tertytchnaya@gmail.com}
}
\keyword{data}
