% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_newsmap_tr}
\alias{data_dictionary_newsmap_tr}
\title{Seed geographical dictionary in Turkish}
\description{
Seed geographical dictionary in Turkish
}
\author{
Lungta Seki \email{yahoo.co.jp0409@gmail.com}
}
\keyword{data}
