% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_newsmap_zh}
\alias{data_dictionary_newsmap_zh}
\title{Seed geographical dictionary in Chinese (simplified)}
\description{
Seed geographical dictionary in Chinese (simplified)

Seed geographical dictionary in Chinese (traditional)
}
\author{
Ke Cheng \email{kecheng.ac@gmail.com}

Chung-hong Chan \email{chainsawtiney@gmail.com}
}
\keyword{data}
