% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_newsmap_zh_cn}
\alias{data_dictionary_newsmap_zh_cn}
\title{Seed geographical dictionary in Chinese (simplified)}
\description{
Seed geographical dictionary in Chinese (simplified)
}
\author{
Ke Cheng \email{kecheng.ac@gmail.com}
}
\keyword{data}
