% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary_newsmap_zh_tw}
\alias{data_dictionary_newsmap_zh_tw}
\title{Seed geographical dictionary in Chinese (traditional)}
\description{
Seed geographical dictionary in Chinese (traditional)
}
\author{
Chung-hong Chan \email{chainsawtiney@gmail.com}
}
\keyword{data}
