% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{findNames}
\alias{findNames}
\title{Identify frequenety capitalized words. Minimum g-socre is 10.83 (p<0.01) by default.}
\usage{
findNames(tokens, count_min, p = 0.001, word_only = TRUE)
}
\arguments{
\item{tokens}{tokenizedTexts object}

\item{count_min}{minimum frequency of names}

\item{p}{p-value for log-likelihood test}

\item{word_only}{return only words if true}
}
\description{
Identify frequenety capitalized words. Minimum g-socre is 10.83 (p<0.01) by default.
}
\examples{
docs <- readLines('/home/kohei/projects/immigration/data/uk_img/2009-2010.txt')
sents <- quanteda::tokenize(docs, what='sentence', simplify = TRUE)
tokens <- quanteda::tokenize(sents, removePunct=TRUE, removeNumbers=TRUE)
names <- findNames(tokens, 5)
}

