% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{findNames2}
\alias{findNames2}
\title{Identify frequenety capitalized words. Minimum g-socre is 10.83 (p<0.01) by default.}
\usage{
findNames2(tokens, count_min, p = 0.001, word_only = TRUE)
}
\arguments{
\item{tokens}{tokenizedTexts object}

\item{count_min}{minimum frequency of names}

\item{p}{p-value for log-likelihood test}

\item{word_only}{return only words if true}
}
\description{
Identify frequenety capitalized words. Minimum g-socre is 10.83 (p<0.01) by default.
}
