% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{findNames3}
\alias{findNames3}
\title{Identify frequenety capitalized words using stopwrods as baseline.}
\usage{
findNames3(tokens, count_min, p = 0.001, word_only = TRUE,
  language = "english")
}
\arguments{
\item{tokens}{tokenizedTexts object}

\item{count_min}{minimum frequency of names}

\item{p}{p-value for log-likelihood test}

\item{word_only}{return only words if true}
}
\description{
Identify frequenety capitalized words using stopwrods as baseline.
}
\examples{
df <- readRDS('/home/kohei/Documents/Newsmap2/yahoo-news.RDS')
today <- as.Date('2016-02-28')
df_recent <- df[today - 7 < df$date & df$date <= today,]
docs <- paste0(df_recent$head, ". ", df_recent$body)
docs <- cleanTexts(docs)
toks <- quanteda::tokenize(docs)
df_nam <- findNames3(toks, count_min=5, word_only=FALSE)
df_nam[order(df_nam$p),]
}
