% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{findNames4}
\alias{findNames4}
\title{Identify frequenety capitalized words with an arbitrary baseline.}
\usage{
findNames4(tokens, count_min, baseline = 0.1, p = 0.001, word_only = TRUE)
}
\arguments{
\item{tokens}{tokenizedTexts object}

\item{count_min}{minimum frequency of names}

\item{p}{p-value for log-likelihood test}

\item{word_only}{return only words if true}
}
\description{
Identify frequenety capitalized words with an arbitrary baseline.
}
