% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{get_country}
\alias{get_country}
\title{Extract most strongly assosiated countries}
\usage{
get_country(pred, rank = c(1, 2))
}
\arguments{
\item{pred}{prediciton by predictCountry}

\item{rank}{rank of country}
}
\description{
Extract most strongly assosiated countries
}

