% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{getCasedTypes}
\alias{getCasedTypes}
\title{Extract unique capitalized tokens}
\usage{
getCasedTypes(tokens, case = "upper")
}
\arguments{
\item{tokens}{tokenizedTexts object}

\item{case}{case of words to be extracted}
}
\description{
Extract unique capitalized tokens
}
