% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{getKeywords}
\alias{getKeywords}
\title{Extract geograpical keywords from a lexicon as sequence of tokens}
\usage{
getKeywords(lexicon, sep = " ")
}
\arguments{
\item{lexicon}{list of place names}

\item{sep_keyword}{separator for keywords in lexicon}
}
\description{
Extract geograpical keywords from a lexicon as sequence of tokens
}

