% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{getTopCountries}
\alias{getTopCountries}
\title{Extract most strongly assosiated countries}
\usage{
getTopCountries(pred, rank = 1)
}
\arguments{
\item{pred}{prediciton by predictCountry}

\item{rank}{rank of country}
}
\description{
Extract most strongly assosiated countries
}

