% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{gscore}
\alias{gscore}
\title{Calculate G-score (it is chi-squre at the moment)}
\usage{
gscore(n_true, n_false, sum_true, sum_false, smooth = 1)
}
\description{
Calculate G-score (it is chi-squre at the moment)
}
