% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{joinNames}
\alias{joinNames}
\title{Idenitfy concatenate sequences of capitalized words}
\usage{
joinNames(tokens, count_min, p = 0.001, verbose = FALSE, types_extra)
}
\arguments{
\item{tokens}{tokenizedTexts object}

\item{count_min}{minimum freqeuency of names}

\item{p}{p-value for log-likelihood test}

\item{verbose}{show progress}

\item{types_extra}{types of tokens that are considered part of names in addition to capitalized words}
}
\description{
Idenitfy concatenate sequences of capitalized words
}
