% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{make_dictionary}
\alias{make_dictionary}
\title{Construct a geographical dictionary from a list of place names}
\usage{
make_dictionary(toks, dict, concatenator = " ", weight = 1, smooth = 1,
  min_count_seq = 2)
}
\arguments{
\item{dict}{quanteda's dictionary object}

\item{smooth}{smoother for coditional frequency}

\item{tokens}{tokenizedTexts object}

\item{sep_keyword}{separator for keywords in lexicon}
}
\description{
Construct a geographical dictionary from a list of place names
}

