% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{makeGeoDictionary}
\alias{makeGeoDictionary}
\title{Construct a geographical dictionary from a list of place names}
\usage{
makeGeoDictionary(tokens, lexicon, power = 1, smooth = 0.001,
  sep_keyword = " ")
}
\arguments{
\item{tokens}{tokenizedTexts object}

\item{lexicon}{list of place names}

\item{power}{weight for word frequency}

\item{smooth}{smoother for coditional frequency}

\item{sep_keyword}{separator for keywords in lexicon}
}
\description{
Construct a geographical dictionary from a list of place names
}
