% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary2.R
\name{makeGeoDictionary2}
\alias{makeGeoDictionary2}
\title{Construct a geographical dictionary from a list of place names}
\usage{
makeGeoDictionary2(tokens, lexicon, power = 1, smooth = 0.001, sep = " ")
}
\arguments{
\item{tokens}{tokenizedTexts object}

\item{lexicon}{list of place names}

\item{smooth}{smoother for coditional frequency}

\item{sep_keyword}{separator for keywords in lexicon}
}
\description{
Construct a geographical dictionary from a list of place names
}
