% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{performance}
\alias{performance}
\title{Evaluate classification perfromance by precision and recall}
\usage{
performance(class_true, class_test)
}
\arguments{
\item{class_true}{vector of true classes}

\item{class_test}{vercor of predicted classes}
}
\description{
Evaluate classification perfromance by precision and recall
}
