% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{predict_country}
\alias{predict_country}
\title{Predict geographical association of texts}
\usage{
predict_country(toks, dict, lang = "english")
}
\arguments{
\item{dict}{geigraphical dicitonary created by makeGeoDictionary}

\item{tokens}{tokenizedTexts}
}
\description{
Predict geographical association of texts
}

