% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{predictCountry}
\alias{predictCountry}
\title{Predict geographical association of texts}
\usage{
predictCountry(tokens, dict)
}
\arguments{
\item{tokens}{tokenizedTexts}

\item{dict}{geigraphical dicitonary created by makeGeoDictionary}
}
\description{
Predict geographical association of texts
}

