% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_newsmap.R
\name{print.textmodel_newsmap_summary}
\alias{print.textmodel_newsmap_summary}
\title{Summary method for fitted Newsmap model}
\usage{
\method{print}{textmodel_newsmap_summary}(object, ...)
}
\arguments{
\item{object}{a fitted Newsmap textmodel}
}
\description{
Summary method for fitted Newsmap model
}
