% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicon.R
\name{readLexicon}
\alias{readLexicon}
\title{Read a list of place names from text a file}
\usage{
readLexicon(file)
}
\arguments{
\item{file}{file path to text file}
}
\description{
Read a list of place names from text a file
}

