% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{removeShortFeatures}
\alias{removeShortFeatures}
\title{Remove short features}
\usage{
removeShortFeatures(tokens, len_min = 3, ...)
}
\arguments{
\item{tokens}{tokenizedTexts object}

\item{len_min}{minimum length of tokens to keep}

\item{...}{passed to quanteda::selectFeatures}
}
\description{
Remove short features
}

