% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{removeSpecialFeatures}
\alias{removeSpecialFeatures}
\title{Select or remove marks and numbers}
\usage{
removeSpecialFeatures(tokens, numbers = TRUE, marks = TRUE, ...)
}
\arguments{
\item{tokens}{tokenizedTexts object}

\item{numbers}{remove numeric features if true}

\item{marks}{remove punctuations and symboles if true}

\item{...}{passed to quanteda::selectFeatures}
}
\description{
Select or remove marks and numbers
}

