% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{selectCasedFeatures}
\alias{selectCasedFeatures}
\title{Select lower or upper-cased words}
\usage{
selectCasedFeatures(tokens, case = "upper", selection = c("keep", "remove"),
  padding = FALSE)
}
\arguments{
\item{tokens}{tokenizedTexts object}

\item{case}{case of words to be extracted}

\item{selection}{keep or remove features}

\item{padding}{leave empty tokens if true}
}
\description{
Select lower or upper-cased words
}
