% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{selectNames}
\alias{selectNames}
\title{Select or remove capitalized words}
\usage{
selectNames(tokens, selection = c("keep", "remove"), padding = FALSE, ...)
}
\arguments{
\item{tokens}{tokenizedTexts object}

\item{selection}{keep or remove features}

\item{padding}{leave empty tokens if true}

\item{...}{passed to quanteda::selectFeatures}
}
\description{
Select or remove capitalized words
}

