% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{stemNames}
\alias{stemNames}
\title{Stem names identified by selectNames}
\usage{
stemNames(names, language = "en", len_min = 5, word_only = TRUE)
}
\arguments{
\item{names}{vector of names identifdied by findNames}

\item{language}{language setting to be passted to the SnowballC stemmer}

\item{len_min}{minimum length of names to be stemmed}

\item{word_only}{return only words if true}
}
\description{
Stem names identified by selectNames
}

