% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{summary.accuracy}
\alias{summary.accuracy}
\title{Print micro and macro average measures of accuracy}
\usage{
\method{summary}{accuracy}(mx)
}
\arguments{
\item{accuracy}{output of accuracy()}
}
\description{
Print micro and macro average measures of accuracy
}

