% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_newsmap.R
\name{summary.textmodel_newsmap}
\alias{summary.textmodel_newsmap}
\title{Summary method for a fitted Newsmap model}
\usage{
\method{summary}{textmodel_newsmap}(object, n = 10, ...)
}
\arguments{
\item{object}{a fitted Newsmap textmodel}

\item{n}{number of classes, features and document names to be shown}

\item{...}{not used.}
}
\description{
Summary method for a fitted Newsmap model
}
