% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_newsmap.R
\name{summary.textmodel_newsmap_accuracy}
\alias{summary.textmodel_newsmap_accuracy}
\title{Print micro and macro average measures of accuracy}
\usage{
\method{summary}{textmodel_newsmap_accuracy}(object, ...)
}
\arguments{
\item{object}{output of accuracy()}

\item{...}{not used.}
}
\description{
Print micro and macro average measures of accuracy
}
