% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_newsmap.R
\name{summary.textmodel_newsmap_fitted}
\alias{summary.textmodel_newsmap_fitted}
\title{Summary method for fitted Newsmap model}
\usage{
\method{summary}{textmodel_newsmap_fitted}(object, n = 10, ...)
}
\arguments{
\item{object}{a fitted Newsmap textmodel}

\item{n}{number of classes, features and document names to be shown}
}
\description{
Summary method for fitted Newsmap model
}
