% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_newsmap.R
\name{accuracy}
\alias{accuracy}
\title{Evaluate classification accuracy in precision and recall}
\usage{
accuracy(x, y)
}
\arguments{
\item{x}{vector of predicted classes}

\item{y}{vector of true classes}
}
\description{
Evaluate classification accuracy in precision and recall
}
\examples{
class_pred <- c('US', 'GB', 'US', 'CN', 'JP', 'FR', 'CN') # prediction
class_true <- c('US', 'FR', 'US', 'CN', 'KP', 'EG', 'US') # true class
acc <- accuracy(class_pred, class_true)
print(acc)
summary(acc)
}
