% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel.R
\name{afe}
\alias{afe}
\title{Compute average feature entropy (AFE)}
\usage{
afe(x, y, smooth = 1)
}
\arguments{
\item{x}{a dfm for features}

\item{y}{a dfm for labels}

\item{smooth}{a numeric value for smoothing to include all the features}
}
\description{
AFE computes randomness of occurrences features in labelled documents.
}
