% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_newsmap.R
\name{print.textmodel_newsmap_summary}
\alias{print.textmodel_newsmap_summary}
\title{Print method for a fitted Newsmap model}
\usage{
\method{print}{textmodel_newsmap_summary}(x, ...)
}
\arguments{
\item{x}{a fitted Newsmap textmodel}

\item{...}{not used.}
}
\description{
Print method for a fitted Newsmap model
}
