% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_newsmap.R
\name{summary.textmodel_newsmap_accuracy}
\alias{summary.textmodel_newsmap_accuracy}
\title{Print micro and macro average measures of accuracy}
\usage{
\method{summary}{textmodel_newsmap_accuracy}(mx)
}
\arguments{
\item{accuracy}{output of accuracy()}
}
\description{
Print micro and macro average measures of accuracy
}
