% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_newsmap.R
\name{summary.textmodel_newsmap_accuracy}
\alias{summary.textmodel_newsmap_accuracy}
\title{Calculate micro and macro average measures of accuracy}
\usage{
\method{summary}{textmodel_newsmap_accuracy}(object, ...)
}
\arguments{
\item{object}{output of accuracy()}

\item{...}{not used.}
}
\description{
This function calculates micro-average precision (p) and recall (r) and
macro-average precision (P) and recall (R) based on a confusion matrix from
\code{accuracy()}.
}
