% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textmodel_newsmap.R
\name{textmodel_newsmap}
\alias{textmodel_newsmap}
\title{supervised multinomial classifier for multinomial text classification}
\usage{
textmodel_newsmap(x, y, smooth = 1,
  verbose = quanteda::quanteda_options("verbose"))
}
\arguments{
\item{x}{dfm from which features will be extracted}

\item{y}{dfm in which features will be class labels}

\item{smooth}{smoothing parameter for word frequency}

\item{verbose}{if \code{TRUE}, show progress of training}
}
\description{
Train a model to predict geographical focus of texts
}
